requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
nextcord==2.6.0
python-dotenv==1.0.0
typing_extensions==4.8.0
yarl==1.9.3
